feat: require authentication

+++ src/http/routes/menu-management/delete-category.ts
@@ -1,5 +1,6 @@
-import { BadRequestError } from '../../errors'
+import { BadRequestError, ForbiddenError } from '../../errors'
 import { FastifyInstance } from 'fastify'
+import { get_auth } from '../../utils/auth'
 import { prisma } from '../../../lib/prisma'
 import { z } from 'zod'
 
@@ -7,6 +8,8 @@ export default async function(app: FastifyInstance) {
   app.delete('/category/:id', async (req, res) => {
     const paramSchema = z.object({ id: z.string().cuid() })
     const { id } = paramSchema.parse(req.params)
+
+    if (!(await get_auth(req, id))) throw new ForbiddenError('No privileges.')
 
     try {
       const category = await prisma.category.delete({ where: { id } })
